+++ source/ModMetadata.cs
@@ -13,7 +13,7 @@ public record ModMetadata : AbstractModMetadata
 
     public override List<string>? Contributors { get; init; }
 
-    public override SemanticVersioning.Version Version { get; init; } = new("1.0.0", false);
+    public override SemanticVersioning.Version Version { get; init; } = new("1.0.1", false);
 
     public override SemanticVersioning.Range SptVersion { get; init; } = new("~4.0.0", false);